% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMarkovTrace.R
\name{CalculateMarkovTrace}
\alias{CalculateMarkovTrace}
\title{A Markov Function}
\usage{
CalculateMarkovTrace(M, p0, n.cycles)
}
\arguments{
\item{M}{Transition probability matrix or array}

\item{p0}{Initial state vector}

\item{n.cycles}{Total number of cycles to run the Markov model}
}
\value{
trace A (n.cycles x n.states) matrix with Markov trace

trans A (n.states x n.states x n.cycles) array with transitions over time
CalculateMarkovTrace()
}
\description{
This function computes the Markov trace and transition array for a given transition Matrix
}
\keyword{Markov}
