% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMicroTrace.R
\name{CalculateMicroTrace}
\alias{CalculateMicroTrace}
\title{CalculateMicroTrace}
\usage{
CalculateMicroTrace(m.Microsim, states, v.Init)
}
\arguments{
\item{m.Microsim}{A microsimulation matrix showing what happens to each individual at each cycle}

\item{v.Init}{Initial state vector}

\item{state_names}{Names of the health states}
}
\value{
trace A(n.t x n.s) matrix with microsimulation trace

trans A (n.s x n.s x n.t) array with transitions over time

trans_ind A (n.s x n.s x n.t x n.i) array with transitions over time per individual

trans_sum_ind A (n.s x n.s x n.i) showing all the transitions during all time points for an individual
CalculateMircoTrace()
}
\description{
This function computes the microsimulation trace and transition array for a given matrix with individual history
}
\keyword{Microsimulation}

