% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEA.R
\name{CEA}
\alias{CEA}
\title{A cost-effectiveness analysis function}
\usage{
CEA(outcomes)
}
\arguments{
\item{outcomes}{Vector with the different rewards}
}
\value{
table Cost effectivess ananlysis results including the ICER.
CEA()
}
\description{
This function performs a cost-effectivenss analysis based on cost and reward data from the simulation.
}
\keyword{Microsimulation}
\keyword{cost-effectiveness}

