% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedValue.R
\name{ExpectedValue}
\alias{ExpectedValue}
\title{A Markov function to compute expected value}
\usage{
ExpectedValue(trans, rwd, r = 0, half = TRUE)
}
\arguments{
\item{trans}{Markov transition array generated by `CalculateMarkovTrace` function}

\item{rwd}{Matrix or array with state rewards (diagonal) and transition rewards}

\item{r}{Discount factor in the scale of a cycle [0, 1]; default = 0}

\item{half}{If TRUE apply Half cycle correction. Defaults to TRUE}
}
\value{
exp.value Expected value
}
\description{
Computes the expected value from a Markov transition array
}
\keyword{Markov}

