% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedValueMicro.R
\name{ExpectedValueMicro}
\alias{ExpectedValueMicro}
\title{ExpectedValueMicro}
\usage{
ExpectedValueMicro(trans, rwd, r = 0, half = TRUE)
}
\arguments{
\item{trans}{A transition array generated by the `CalculateMicroTrace` function}

\item{rwd}{List of matrices with state rewards (diagonal) and transition rewards (off-diagonal)}

\item{r}{Discount factor; default = 0.}

\item{half}{If TRUE apply Half cycle correction}
}
\value{
exp_value_ind Expected value per individual for the different rewards in the list

mean Mean Expeced value for the different rewards in the list
ExpectedValueMicro()
}
\description{
This function computes the expected value per individuals from a microsimulation transition array
}
\keyword{costs}
\keyword{outcomes}
\keyword{qaly}
\keyword{rewards}

