% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedValueMicroTimeDepReward.R
\name{ExpectedValueMicroTimeDepReward}
\alias{ExpectedValueMicroTimeDepReward}
\title{ExpectedValueMicroTimeDepReward}
\usage{
ExpectedValueMicroTimeDepReward(trans, cum, rwd, r = 0, half = TRUE)
}
\arguments{
\item{trans}{A transition array generated by the `CalculateMicroTrace` function}

\item{cum}{A matrix with the cummulative time being sick or sicker}

\item{rwd}{List of matrices with state rewards (diagonal) and transition rewards (off-diagonal)}

\item{r}{Discount factor; default = 0.}

\item{half}{If TRUE apply Half cycle correction}
}
\value{
exp_value_ind Expected value per individual for the different rewards in the list

mean Mean expected value for the different rewards in the list
ExpectedValueMicroTimeDepReward()
}
\description{
This function computes the expected value per individuals from a Microsimulation transition array. While running the function treatment utility decreases with increasing years of being sick. Note: only usefull for the Sick-Sicker model.
}
\keyword{Microsimulation}
\keyword{costs}
\keyword{outcomes}
\keyword{qaly}
\keyword{rewards}

