% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{\code{ggplot2}-like colour scale in HCL space}
\usage{
gg_color_hue(n, hue_min = 15, hue_max = 375, l = 65, c = 100)
}
\arguments{
\item{n}{Number of colours to return}

\item{hue_min}{Minimum hue value in the range [0, 360].}

\item{hue_max}{Maximum hue value in the range [0, 360].}

\item{l}{Luminance in the range [0, 100].}

\item{c}{Chroma of the colour.}
}
\description{
Function for number of ticks in axis of \code{ggplot2} plots.
}
\section{Details}{

}

\examples{
gg_color_hue(10)
}
\keyword{ggplot2}
