% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroSim.R
\name{MicroSim}
\alias{MicroSim}
\title{A Microsimulation function}
\usage{
MicroSim(m.Transitions, p.0, n.i, n.t, seed = NULL)
}
\arguments{
\item{m.Transitions}{Transition probability matrix (2D, 3D or 4D)}

\item{p.0}{Initial state vector}

\item{n.i}{Number of hypothetical individuals simulated in the model}

\item{n.t}{Total number of cycles to run the model}

\item{seed}{Seed number for random number generator (optional)}
}
\value{
m.Microsim A (n.i x n.t) matrix showing state history of each individual at each cycle
MicroSim()
}
\description{
This function computes a microsimulation matrix that keeps track of what happens to each individual at each cycle
}
\keyword{microsimulation}

