% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroSimDeath.R
\name{MicroSimDeath}
\alias{MicroSimDeath}
\title{MicroSimDeath}
\usage{
MicroSimDeath(a.Transitions, p.0, n.i, n.t, seed = NULL, cum = FALSE)
}
\arguments{
\item{a.Transitions}{Transition probability array  (n.s x n.s x n.t x n.i)}

\item{p.0}{Initial state vector}

\item{n.i}{Number of hypothetical individuals simulated in the model}

\item{n.t}{Total number of cycles to run the model}

\item{seed}{Seed number for random number generator (optional)}

\item{cum}{Returns the cumulative variable that keeps track of the time spent as being sick (default = FALSE)}
}
\value{
m.Microsim A (n.i x n.t) matrix showing what happens to each individual at each cycle

cum A (n.i x n.t) matrix showing how long an individual has been ill
MicroSimDeath()
}
\description{
This function computes the microsimulation matrix keeping track of what happens to each individual at each cycle. While doing this the function includes age dependent mortality and has the opportunity to give the variable cumulative sick as output. This function is not universal and should only be used for the Sick-Sicker model example
}
\keyword{Microsimulation}

