% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl.R
\name{mnl}
\alias{mnl}
\title{Multinomial Logistic (MNL) Metamodeling for Decision Sensitivity Analysis (DSA)}
\usage{
mnl(x, y, parm, strategies = NULL, range = NULL, txtsize = 12)
}
\arguments{
\item{x}{Matrix with the model inputs or parameters}

\item{y}{Matrix with the model outputs}

\item{parm}{String with the name of the parameter of interest}

\item{strategies}{String vector with the name of the strategies}

\item{range}{Range of the parameter of interest. Default: NULL range. If
range=NULL, the 2.5th and 9.75th percentile of the parameter
of interest will be used as lower and upper bounds of the 
range, respectively.}

\item{txtsize}{Font size for ggplot graph. Default: 12}
}
\value{
mnl A melted data frame with each strategy's probability of being 
cost-effective over a range of values of the parameter of interest.
}
\description{
\code{mnl} is used to compute the probability of each startegy being optimal
over different values of a parametere of interest from a probabilistic 
sensitivity analysis (PSA) dataset.
}
\section{Details}{

\code{mnl} computes the probability of each of the strategies being 
cost-effective at over different values of the parameter of interest.
}

\keyword{analysis;}
\keyword{decision}
\keyword{logistic}
\keyword{metamodel}
\keyword{multinomial}
\keyword{sensitivity}
