% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeCI.R
\name{OutcomeCI}
\alias{OutcomeCI}
\title{Outcome uncertainty}
\usage{
OutcomeCI(outcomes)
}
\arguments{
\item{outcomes}{Vector with the different rewards}
}
\value{
table with the mean outcome values and the upper and lower value of the 95% confidence interval
CEA()
}
\description{
This function created 95% confidence intervals around the outcome values
}
\keyword{Microsimulation}
\keyword{outcomes}
\keyword{uncertainty}

