% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.evpi.R
\name{plot.evpi}
\alias{plot.evpi}
\title{Plot of Expected Value of Perfect Information (EVPI)}
\usage{
\method{plot}{evpi}(evpi, title = "Expected Value of Perfect Information",
  txtsize = 12, currency = "$")
}
\arguments{
\item{evpi}{Object of class \code{evpi}. A data frame produced by function 
\code{evpi} with the EVPI for each willingness-to-pay (WTP) threshold}

\item{title}{String with graph's title}

\item{txtsize}{number with text size}

\item{currency}{String with currency used in the cost-effectiveness analysis (CEA). 
Default: $, but it could be any currency symbol or word (e.g., £, €, peso)}
}
\value{
plot.gg A \code{ggplot2} object with the EVPI
}
\description{
Plots the EVPI as a \code{ggplot2} object calculated with \code{\link{evpi}}.
}
\keyword{expected}
\keyword{information}
\keyword{of}
\keyword{perfect}
\keyword{value}
