% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMarkovTrace.R
\name{PlotTrace}
\alias{PlotTrace}
\title{A plot Function}
\usage{
PlotTrace(trace, title, xlab, txtsize = 12)
}
\arguments{
\item{trace}{Trace generated by 'CalculateMarkovTrace' or 'CalculateMicroTrace'}

\item{title}{Title of the plot}

\item{xlab}{x-axis label}

\item{txtsize}{Text size for plot, default = 12}

}
\value{
plot_trace ggplot of the trace from the Markov of Micro-simulation

PlotTrace()
}
\description{
This function generates a plot from the trace generated by the functions 'CalculateMarkovTrace' or 'CalculateMicroTrace'. 
}
\keyword{Markov}
\keyword{Microsimulation"
\keyword{plot}

