% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbToOdds.R
\name{ProbToOdds}
\alias{ProbToOdds}
\title{A probability function}
\usage{
ProbToOdds(p)
}
\arguments{
\item{p}{The probability}
}
\value{
The odds for the given probability.

ProbToOdds()
}
\description{
This function converts a probability into a odds
}
\keyword{odds}
\keyword{probability}

