% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateToProb.R
\name{RateToProb}
\alias{RateToProb}
\title{A bas}
\usage{
RateToProb(r, t)
}
\arguments{
\item{r}{Rate}

\item{t}{Time}
}
\value{
Probability for the time interval similar to the rate
RateToProb()
}
\description{
This function computes the microsimulation trace and transition array for a given matrix with individual history
}
\keyword{probability}
\keyword{rate}

