% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrace.R
\name{PlotTrace}
\alias{PlotTrace}
\title{A plot function}
\usage{
PlotTrace(trace, title, xlab, txtsize = 12)
}
\arguments{
\item{trace}{Trace generated by `CalculateMarkovTrace` function or "CalculateMicroTrace"}

\item{title}{Title of the plot, e.g. "Markov Trace" or "Microsimulation Trace"}

\item{xlab}{x-axis label}

\item{txtsize}{Text size for plot, default = 12}
}
\value{
plot_trace ggplot of the Markov trace  
PlotTrace()
}
\description{
This function generates a graph from the Markov or Micro Trace 
The graph shows distribution of the cohort (of inidvidual subjects) over the different health states per cycle.
}
\keyword{Markov;}
\keyword{graph}
\keyword{plot;}

